Extract self-contained enum
